pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tender-extract"
version = "0.1.0"
description = "面向中文标书的混合抽取流水线"
authors = [
    {name = "Tender Extract Team", email = "team@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "markdown-it-py>=3.0.0",
    "pyahocorasick>=2.0.0",
    "rapidfuzz>=3.0.0",
    "datasketch>=1.5.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "tqdm>=4.65.0",
    "openai>=1.0.0",
    "ollama>=0.1.0",
    "jieba>=0.42.1",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "langchain-text-splitters>=0.3.0",
]

[project.optional-dependencies]
cli = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "tqdm>=4.65.0",
]
ner = [
    "jieba>=0.42.1",
]

[project.scripts]
tender-extract = "tender_extract.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/tender_extract"]

[tool.typer]
help = "面向中文标书的混合抽取流水线" 